+++ Changelog.md
@@ -14,6 +14,7 @@
 * SDK bump - Fixes CXV1-11151
 * CXV1-9144 - Toolbar view of wrapups
 * CXV1-11049 - Remove extraneous /queues calls, add excludeOffline to TransferMenu /users call
+* CXV1-11131 - Fix stats issues with localisation
 
 ### [1.1.0] (2017-07-21)
 * BUGFIX - SDK version bump to 5.3.27-SNAPSHOT